Week2/requirements.txt
ackermann-msgs==2.0.2
action-msgs==2.0.2
action-tutorials-interfaces==0.33.5
action-tutorials-py==0.33.5
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
ament-cmake-test==2.5.4
ament-copyright==0.17.2
ament-cppcheck==0.17.2
ament-cpplint==0.17.2
ament-flake8==0.17.2
ament-index-python==1.8.1
ament-lint==0.17.2
ament-lint-cmake==0.17.2
ament-package==0.16.4
ament-pep257==0.17.2
ament-uncrustify==0.17.2
ament-xmllint==0.17.2
angles==1.16.0
blinker==1.9.0
builtin-interfaces==2.0.2
click==8.2.1
composition-interfaces==2.0.2
control-msgs==5.4.0
controller-manager==4.29.0
controller-manager-msgs==4.29.0
cv-bridge==4.1.0
demo-nodes-py==0.33.5
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.4
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.5
examples-rclpy-minimal-action-client==0.19.5
examples-rclpy-minimal-action-server==0.19.5
examples-rclpy-minimal-client==0.19.5
examples-rclpy-minimal-publisher==0.19.5
examples-rclpy-minimal-service==0.19.5
examples-rclpy-minimal-subscriber==0.19.5
Flask==3.1.1
generate-parameter-library-py==0.4.0
geometry-msgs==5.3.6
gps-msgs==2.0.4
image-geometry==4.1.0
interactive-markers==2.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
laser-geometry==2.7.0
launch==3.4.4
launch-ros==0.26.7
launch-testing==3.4.4
launch-testing-ros==0.26.7
launch-xml==3.4.4
launch-yaml==3.4.4
lifecycle-msgs==2.0.2
logging-demo==0.33.5
map-msgs==2.4.1
MarkupSafe==3.0.2
message-filters==4.11.6
nav-msgs==5.3.6
osrf-pycommon==2.1.6
pal-statistics==2.6.2
pal-statistics-msgs==2.6.2
pcl-msgs==1.0.0
pendulum-msgs==0.33.5
pyserial==3.5
python-qt-binding==2.2.1
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.5
rcl-interfaces==2.0.2
rclpy==7.1.4
rcutils==6.7.2
resource-retriever==3.4.4
rmw-dds-common==3.1.0
ros-gz-bridge==1.0.12
ros-gz-interfaces==1.0.12
ros-gz-sim==1.0.12
ros2action==0.32.4
ros2bag==0.26.7
ros2bag-mcap-cli==0.26.7
ros2bag-sqlite3-cli==0.26.7
ros2cli==0.32.4
ros2component==0.32.4
ros2controlcli==4.29.0
ros2doctor==0.32.4
ros2interface==0.32.4
ros2launch==0.26.7
ros2lifecycle==0.32.4
ros2multicast==0.32.4
ros2node==0.32.4
ros2param==0.32.4
ros2pkg==0.32.4
ros2run==0.32.4
ros2service==0.32.4
ros2topic==0.32.4
rosbag2-interfaces==0.26.7
rosbag2-py==0.26.7
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.5
rosidl-cli==4.6.5
rosidl-cmake==4.6.5
rosidl-generator-c==4.6.5
rosidl-generator-cpp==4.6.5
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.5
rosidl-parser==4.6.5
rosidl-pycommon==4.6.5
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.1
rosidl-typesupport-fastrtps-cpp==3.6.1
rosidl-typesupport-introspection-c==4.6.5
rosidl-typesupport-introspection-cpp==4.6.5
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.4
rqt-bag-plugins==1.5.4
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.3
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.2
shape-msgs==5.3.6
sros2==0.13.3
statistics-msgs==2.0.2
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.36.11
tf2-kdl==0.36.11
tf2-msgs==0.36.11
tf2-py==0.36.11
tf2-ros-py==0.36.11
tf2-sensor-msgs==0.36.11
tf2-tools==0.36.11
theora-image-transport==4.0.4
topic-monitor==0.33.5
trajectory-msgs==5.3.6
turtlebot3-example==2.1.5
turtlebot3-msgs==2.2.3
turtlebot3-teleop==2.1.5
turtlesim==1.8.3
type-description-interfaces==2.0.2
unique-identifier-msgs==2.5.0
vision-msgs==4.1.1
visualization-msgs==5.3.6
Werkzeug==3.1.3
xacro==2.0.13
